refactor: remove local-storage in favor of the new dependency
